create new ibc/types dir
